refactor: extract common findValidPort method
