+++ fe/fe-core/src/test/java/org/apache/doris/journal/bdbje/BDBEnvironmentTest.java
@@ -21,6 +21,7 @@ import org.apache.doris.catalog.Env;
 import org.apache.doris.common.Config;
 import org.apache.doris.common.Pair;
 import org.apache.doris.common.jmockit.Deencapsulation;
+import org.apache.doris.common.util.UnitTestUtil;
 import org.apache.doris.ha.FrontendNodeType;
 import org.apache.doris.system.Frontend;
 
@@ -47,13 +48,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.RepeatedTest;
 
 import java.io.File;
-import java.io.IOException;
 import java.lang.reflect.Field;
-import java.net.DatagramSocket;
-import java.net.InetSocketAddress;
-import java.net.ServerSocket;
-import java.net.Socket;
-import java.net.SocketException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -61,7 +56,6 @@ import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import java.util.concurrent.CountDownLatch;
 
 
 public class BDBEnvironmentTest {
@@ -102,44 +96,6 @@ public class BDBEnvironmentTest {
         }
     }
 
-    private int findValidPort() {
-        int port = 0;
-        for (int i = 0; i < 65535; i++) {
-            try (ServerSocket socket = new ServerSocket(0)) {
-                socket.setReuseAddress(true);
-                port = socket.getLocalPort();
-                try (DatagramSocket datagramSocket = new DatagramSocket(port)) {
-                    datagramSocket.setReuseAddress(true);
-                    break;
-                } catch (SocketException e) {
-                    LOG.info("The port {} is invalid and try another port", port);
-                }
-
-                CountDownLatch latch = new CountDownLatch(1);
-                final int serverPort = port;
-                new Thread(() -> {
-                    try (Socket clientSocket = new Socket()) {
-                        clientSocket.connect(new InetSocketAddress("localhost", serverPort));
-                        latch.await();  // Wait until the server closes the connection
-                    } catch (IOException | InterruptedException e) {
-                        // CHECKSTYLE IGNORE THIS LINE
-                    }
-                }).start();
-                // Accept a connection from the client
-                try (Socket serverConn = socket.accept()) {
-                    // CHECKSTYLE IGNORE THIS LINE
-                } catch (IOException e) {
-                    // CHECKSTYLE IGNORE THIS LINE
-                }
-                latch.countDown();  // Signal that the server has closed the connection
-            } catch (IOException e) {
-                throw new IllegalStateException("Could not find a free TCP/IP port");
-            }
-        }
-        Preconditions.checkArgument(((port > 0) && (port < 65536)));
-        return port;
-    }
-
     private byte[] randomBytes() {
         byte[] byteArray = new byte[32];
         new SecureRandom().nextBytes(byteArray);
@@ -149,7 +105,7 @@ public class BDBEnvironmentTest {
     // @Test
     @RepeatedTest(1)
     public void testSetup() throws Exception {
-        int port = findValidPort();
+        int port = UnitTestUtil.findValidPort();
         String selfNodeName = Env.genFeNodeName("127.0.0.1", port, false);
         String selfNodeHostPort = "127.0.0.1:" + port;
         if (LOG.isDebugEnabled()) {
@@ -228,7 +184,7 @@ public class BDBEnvironmentTest {
     // @Test
     @RepeatedTest(1)
     public void testSetupTwice() throws Exception {
-        int port = findValidPort();
+        int port = UnitTestUtil.findValidPort();
         String selfNodeName = Env.genFeNodeName("127.0.0.1", port, false);
         String selfNodeHostPort = "127.0.0.1:" + port;
         File homeFile = new File(createTmpDir());
@@ -242,7 +198,7 @@ public class BDBEnvironmentTest {
     // @Test
     @RepeatedTest(1)
     public void testMetadataRecovery() throws Exception {
-        int port = findValidPort();
+        int port = UnitTestUtil.findValidPort();
         String selfNodeName = Env.genFeNodeName("127.0.0.1", port, false);
         String selfNodeHostPort = "127.0.0.1:" + port;
 
@@ -276,7 +232,7 @@ public class BDBEnvironmentTest {
     // @Test
     @RepeatedTest(1)
     public void testOpenReplicatedEnvironmentTwice() throws Exception {
-        int port = findValidPort();
+        int port = UnitTestUtil.findValidPort();
         String selfNodeName = Env.genFeNodeName("127.0.0.1", port, false);
         String selfNodeHostPort = "127.0.0.1:" + port;
 
@@ -312,7 +268,7 @@ public class BDBEnvironmentTest {
     // @Test
     @RepeatedTest(1)
     public void testCluster() throws Exception {
-        int masterPort = findValidPort();
+        int masterPort = UnitTestUtil.findValidPort();
         String masterNodeName = Env.genFeNodeName("127.0.0.1", masterPort, false);
         String masterNodeHostPort = "127.0.0.1:" + masterPort;
         if (LOG.isDebugEnabled()) {
@@ -326,7 +282,7 @@ public class BDBEnvironmentTest {
         List<BDBEnvironment> followerEnvironments = new ArrayList<>();
         List<File> followerDirs = new ArrayList<>();
         for (int i = 1; i <= 2; i++) {
-            int followerPort = findValidPort();
+            int followerPort = UnitTestUtil.findValidPort();
             String followerNodeName = Env.genFeNodeName("127.0.0.1", followerPort, false);
             String followerNodeHostPort = "127.0.0.1:" + followerPort;
             if (LOG.isDebugEnabled()) {
@@ -340,7 +296,7 @@ public class BDBEnvironmentTest {
             followerEnvironments.add(followerEnvironment);
         }
 
-        int observerPort = findValidPort();
+        int observerPort = UnitTestUtil.findValidPort();
         String observerNodeName = Env.genFeNodeName("127.0.0.1", observerPort, false);
         String observerNodeHostPort = "127.0.0.1:" + observerPort;
         if (LOG.isDebugEnabled()) {
@@ -437,7 +393,7 @@ public class BDBEnvironmentTest {
         LOG.info("start");
         List<Pair<BDBEnvironment, NodeInfo>> followersInfo = new ArrayList<>();
 
-        int masterPort = findValidPort();
+        int masterPort = UnitTestUtil.findValidPort();
         String masterNodeName = "fe1";
         String masterNodeHostPort = "127.0.0.1:" + masterPort;
 
@@ -447,7 +403,7 @@ public class BDBEnvironmentTest {
         followersInfo.add(Pair.of(masterEnvironment, new NodeInfo(masterNodeName, masterNodeHostPort, masterDir)));
 
         for (int i = 2; i <= 3; i++) {
-            int nodePort = findValidPort();
+            int nodePort = UnitTestUtil.findValidPort();
             String nodeName = "fe" + i;
             String nodeHostPort = "127.0.0.1:" + nodePort;
 
@@ -606,7 +562,7 @@ public class BDBEnvironmentTest {
     public void testReadTxnIsNotMatched() throws Exception {
         List<Pair<BDBEnvironment, NodeInfo>> followersInfo = new ArrayList<>();
 
-        int masterPort = findValidPort();
+        int masterPort = UnitTestUtil.findValidPort();
         String masterNodeName = "fe1";
         String masterNodeHostPort = "127.0.0.1:" + masterPort;
 
@@ -616,7 +572,7 @@ public class BDBEnvironmentTest {
         followersInfo.add(Pair.of(masterEnvironment, new NodeInfo(masterNodeName, masterNodeHostPort, masterDir)));
 
         for (int i = 2; i <= 3; i++) {
-            int nodePort = findValidPort();
+            int nodePort = UnitTestUtil.findValidPort();
             String nodeName = "fe" + i;
             String nodeHostPort = "127.0.0.1:" + nodePort;
