+++ fe/fe-core/src/test/java/org/apache/doris/utframe/TestWithFeService.java
@@ -56,6 +56,7 @@ import org.apache.doris.common.DdlException;
 import org.apache.doris.common.FeConstants;
 import org.apache.doris.common.MetaNotFoundException;
 import org.apache.doris.common.util.SqlParserUtils;
+import org.apache.doris.common.util.UnitTestUtil;
 import org.apache.doris.datasource.CatalogIf;
 import org.apache.doris.nereids.CascadesContext;
 import org.apache.doris.nereids.StatementContext;
@@ -102,11 +103,6 @@ import org.junit.jupiter.api.TestInstance;
 import java.io.File;
 import java.io.IOException;
 import java.io.StringReader;
-import java.net.DatagramSocket;
-import java.net.InetSocketAddress;
-import java.net.ServerSocket;
-import java.net.Socket;
-import java.net.SocketException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -115,7 +111,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
-import java.util.concurrent.CountDownLatch;
 
 /**
  * This is the base class for unit class that wants to start a FE service.
@@ -383,11 +378,11 @@ public abstract class TestWithFeService {
         }
         System.out.println("CREATE FE SERVER DIR: " + Config.custom_config_dir);
 
-        int feHttpPort = findValidPort();
-        int feRpcPort = findValidPort();
-        int feQueryPort = findValidPort();
-        int arrowFlightSqlPort = findValidPort();
-        int feEditLogPort = findValidPort();
+        int feHttpPort = UnitTestUtil.findValidPort();
+        int feRpcPort = UnitTestUtil.findValidPort();
+        int feQueryPort = UnitTestUtil.findValidPort();
+        int arrowFlightSqlPort = UnitTestUtil.findValidPort();
+        int feEditLogPort = UnitTestUtil.findValidPort();
         Map<String, String> feConfMap = Maps.newHashMap();
         // set additional fe config
         feConfMap.put("http_port", String.valueOf(feHttpPort));
@@ -474,11 +469,11 @@ public abstract class TestWithFeService {
     }
 
     private Backend createBackendWithoutRetry(String beHost, int feRpcPort) throws IOException, InterruptedException {
-        int beHeartbeatPort = findValidPort();
-        int beThriftPort = findValidPort();
-        int beBrpcPort = findValidPort();
-        int beHttpPort = findValidPort();
-        int beArrowFlightSqlPort = findValidPort();
+        int beHeartbeatPort = UnitTestUtil.findValidPort();
+        int beThriftPort = UnitTestUtil.findValidPort();
+        int beBrpcPort = UnitTestUtil.findValidPort();
+        int beHttpPort = UnitTestUtil.findValidPort();
+        int beArrowFlightSqlPort = UnitTestUtil.findValidPort();
 
         // start be
         MockedBackendFactory.BeThriftService beThriftService = new DefaultBeThriftServiceImpl();
@@ -516,43 +511,6 @@ public abstract class TestWithFeService {
         } catch (IOException e) {
             e.printStackTrace();
         }
-    }
-
-    public static int findValidPort() {
-        int port = 0;
-        while (true) {
-            try (ServerSocket socket = new ServerSocket(0)) {
-                socket.setReuseAddress(true);
-                port = socket.getLocalPort();
-                try (DatagramSocket datagramSocket = new DatagramSocket(port)) {
-                    datagramSocket.setReuseAddress(true);
-                    break;
-                } catch (SocketException e) {
-                    System.out.println("The port " + port + " is invalid and try another port.");
-                }
-
-                CountDownLatch latch = new CountDownLatch(1);
-                final int serverPort = port;
-                new Thread(() -> {
-                    try (Socket clientSocket = new Socket()) {
-                        clientSocket.connect(new InetSocketAddress("localhost", serverPort));
-                        latch.await();  // Wait until the server closes the connection
-                    } catch (IOException | InterruptedException e) {
-                        // CHECKSTYLE IGNORE THIS LINE
-                    }
-                }).start();
-                // Accept a connection from the client
-                try (Socket serverConn = socket.accept()) {
-                    // CHECKSTYLE IGNORE THIS LINE
-                } catch (IOException e) {
-                    // CHECKSTYLE IGNORE THIS LINE
-                }
-                latch.countDown();  // Signal that the server has closed the connection
-            } catch (IOException e) {
-                throw new IllegalStateException("Could not find a free TCP/IP port to start HTTP Server on");
-            }
-        }
-        return port;
     }
 
     public String getSQLPlanOrErrorMsg(String sql) throws Exception {
